# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: SmartTrainer API
  description: |
    Spring Boot 製トレーニング管理アプリケーションの API 仕様書。
    認証は Spring Security によるセッションベースのフォームログインおよび
    OAuth2（Google / LINE）に対応。
    管理者専用エンドポイントは ADMIN ロールが必要。
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
...
paths:
...
  /user/body:
    get:
      tags: [body-measurement]
      summary: 体重記録画面表示
      description: |
        ログインユーザーの体重記録一覧とChart.js折れ線グラフを表示する。認証必須。
        記録一覧（日付・体重・体脂肪率・メモ）＋入力フォーム＋グラフで構成。
        グラフデータはサーバーサイドで生成して `<script>` タグに埋め込む。
      security:
        - sessionAuth: []
      responses:
        "200":
          description: 体重記録画面 HTML（一覧・グラフ・入力フォーム）
        "302":
          description: 未認証 → /login へリダイレクト
  /api/body:
    post:
      tags: [body-measurement]
      summary: 体重データ保存（新規登録・上書き）
      description: |
        指定日付の体重データを保存する。認証必須。
        同一日付が既に存在する場合は上書き（UPSERT）。
        userId はサーバーサイドで解決（IDOR防止）。
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [measuredDate, weightKg]
              properties:
                measuredDate:
                  type: string
                  format: date
                  description: 計測日
                  example: "2026-06-06"
                weightKg:
                  type: number
                  format: double
                  description: 体重（kg）。20.0〜300.0の範囲
                  example: 72.5
                bodyFatPct:
                  type: number
                  format: double
                  description: 体脂肪率（%）。0.0〜60.0の範囲。任意
                  example: 18.5
                  nullable: true
                memo:
                  type: string
                  description: メモ（200文字以内）。任意
                  maxLength: 200
                  nullable: true
      responses:
        "200":
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: ok
        "400":
          description: バリデーションエラー（体重未入力・範囲外・日付不正）
        "302":
          description: 未認証 → /login へリダイレクト
        "429":
          $ref: "./schemas/responses/TooManyRequestsResponse.yaml"
  /api/body/{id}:
    delete:
      tags: [body-measurement]
      summary: 体重データ削除
      description: |
        指定IDの体重記録を削除する。所有者チェックあり。認証必須。
        userId はサーバーサイドで解決（IDOR防止）。
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 体重記録 ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: deleted
        "403":
          description: 他ユーザーの記録へのアクセス
        "404":
          description: 指定記録が存在しない
        "302":
          description: 未認証 → /login へリダイレクト
        "429":
          $ref: "./schemas/responses/TooManyRequestsResponse.yaml"

  # ── メールアドレス変更 ──────────────────────────────────────────────
  /user/email:
    get:
      tags: [email-change]
      summary: メールアドレス変更フォーム表示
      description: |
        ログインユーザーのメールアドレス変更フォームを表示する。認証必須。
        Google/LINE OAuth2 ユーザー（password=NULL）は変更不可。
        UI に「OAuth2ログインユーザーはメールアドレスを変更できません」と表示する。
      security:
        - sessionAuth: []
      responses:
        "200":
          description: |
            メールアドレス変更フォーム HTML。
            パスワードログインユーザー: 新メールアドレス + 現パスワード入力フォームを表示。
            OAuth2ユーザー: 変更不可メッセージを表示（フォームなし）。
        "302":
          description: 未認証 → /login へリダイレクト
    post:
      tags: [email-change]
      summary: メールアドレス変更申請
      description: |
        新メールアドレスへの確認メールを送信する。認証必須。パスワードログインユーザーのみ。

        **処理フロー**:
        1. 現在のパスワードを照合（本人確認）
        2. 新メールアドレスの重複チェック
        3. トークンを生成し `email_change_tokens` テーブルに保存（有効期限24時間）
        4. 新メールアドレスへ確認メールを送信

        **エラー**:
        - 現パスワード不一致 → フォーム再表示
        - 新メールアドレス重複 → フォーム再表示
        - OAuth2ユーザー → 400
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [newEmail, currentPassword]
              properties:
                newEmail:
                  type: string
                  format: email
                  description: 新しいメールアドレス
                  maxLength: 255
                currentPassword:
                  type: string
                  format: password
                  description: 現在のパスワード（本人確認用）
      responses:
        "302":
          description: 申請完了（確認メール送信済み） → /user/email?sent=true へリダイレクト
        "200":
          description: バリデーションエラー（パスワード不一致・メール重複）→ フォーム再表示
        "400":
          description: OAuth2ユーザーによるアクセス（変更不可）
        "302 (unauthenticated)":
          description: 未認証 → /login へリダイレクト
  /user/email/confirm:
    get:
      tags: [email-change]
      summary: メールアドレス変更確認（トークン検証・確定）
      description: |
        確認メールのリンクをクリックした際にアクセスされるエンドポイント。
        トークンを検証し、メールアドレスを更新する。認証必須。

        **処理フロー**:
        1. トークンの存在確認
        2. 有効期限チェック（24時間）
        3. 新メールアドレスの重複再チェック（他ユーザーが先に同じアドレスを取得していないか）
        4. users.email を新メールアドレスに更新
        5. email_change_tokens から削除

        **エラー**:
        - トークン不存在・期限切れ → エラーページ表示
        - メール重複 → エラーページ表示
      security:
        - sessionAuth: []
      parameters:
        - name: token
          in: query
          required: true
          description: 確認メールに含まれるトークン（UUID）
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "302":
          description: 変更成功 → /user/email?changed=true へリダイレクト
        "200":
          description: トークン無効・期限切れ・メール重複 → エラーメッセージ付きページ表示
        "302 (unauthenticated)":
          description: 未認証 → /login へリダイレクト

# ──────────────────────────────────────────────────────────────────
# Components（セキュリティスキームのみ。スキーマは schemas/ を参照）
# ──────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: |
        Spring Security のセッション認証。
        ログイン後に JSESSIONID クッキーが付与される。